requirements.txt
langchain-core
groq
langchain-community
python-dotenv